splitting our code, offering better separation of concerns and modularity
