+++ 19---diving-into-redux-an-alternative-to-the-context-API/01-starting-project/src/components/Auth.js
@@ -1,6 +1,6 @@
 import classes from './Auth.module.css';
 import { useDispatch } from 'react-redux';
-import { authActions } from '../store';
+import { authActions } from '../store/auth';
 
 const Auth = () => {
