+++ 19---diving-into-redux-an-alternative-to-the-context-API/01-starting-project/src/components/Counter.js
@@ -1,12 +1,13 @@
 import { useSelector, useDispatch } from 'react-redux';
 import classes from './Counter.module.css';
-import { counterActions } from '../store/index';
+import { counterActions } from '../store/counter';
 
 const Counter = () => {
 
   const counter = useSelector((state) => state.counter.value);
   const show = useSelector((state) => state.counter.show);
   const dispatch = useDispatch();
+
 
   const incrementHandler = () => {
     dispatch(counterActions.increment());
