+++ 19---diving-into-redux-an-alternative-to-the-context-API/01-starting-project/src/components/Header.js
@@ -1,6 +1,6 @@
 import classes from './Header.module.css';
 import { useSelector } from 'react-redux';
-import { authActions } from '../store';
+import { authActions } from '../store/auth';
 import { useDispatch } from 'react-redux';
 
 const Header = () => {
