+++ 19---diving-into-redux-an-alternative-to-the-context-API/01-starting-project/src/store/counter.js
@@ -1,0 +1,28 @@
+import { createSlice } from '@reduxjs/toolkit';
+
+const initialCounterState = {
+    value: 0,
+    show: true
+}
+
+const counterSlice = createSlice({
+    name: 'counter',
+    initialState: initialCounterState,
+    reducers: {
+        increment(state) {
+            state.value++;
+        },
+        decrement(state) {
+            state.value--;
+        },
+        increase(state, action) {
+            state.value = state.value + action.payload;
+        },
+        toggleCounter(state) {
+            state.show = !state.show;
+        }
+    }
+})
+
+export default counterSlice.reducer;
+export const counterActions = counterSlice.actions;
